# docs/swagger.yaml
basePath: /
definitions:
...
  me.updateMyPasswordForm:
    properties:
      confirm_password:
        type: string
      new_password:
        type: string
      old_password:
        type: string
      verification_code:
        type: string
    type: object
...
info:
  contact: {}
  description: Proman Backend API
  title: Proman Backend
paths:
...
  /api/me:
    get:
      consumes:
      - application/json
      operationId: my-profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my info
      tags:
      - Me
    put:
      consumes:
      - application/json
      operationId: update-my-profile
      parameters:
      - description: name
        in: formData
        name: name
        type: string
      - description: email
        in: formData
        name: email
        type: string
      - description: position
        in: formData
        name: position
        type: string
      - description: phone
        in: formData
        name: phone
        type: string
      - description: avatar
        in: formData
        name: avatar
        type: file
      - description: verification_code
        in: formData
        name: verification_code
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update my profile
      tags:
      - Me
  /api/me/password:
    put:
      consumes:
      - application/json
      operationId: update-my-password
      parameters:
      - description: update my password json
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/me.updateMyPasswordForm'
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Update my password
      tags:
      - Me
  /api/me/project/count:
    get:
      consumes:
      - application/json
      operationId: my-project-count
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my project count
      tags:
      - Me Project
  /api/me/project/count/type:
    get:
      consumes:
      - application/json
      operationId: my-project-count-type
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my project count by type
      tags:
      - Me Project
  /api/me/projects:
    get:
      consumes:
      - application/json
      operationId: my-projects
      parameters:
      - description: Search by nama or description
        in: query
        name: q
        type: string
      - description: Search by status
        enum:
        - active
        - completed
        - pending
        - cancelled
        in: query
        name: status
        type: string
      - description: Start date
        in: query
        name: start
        type: string
      - description: End date
        in: query
        name: end
        type: string
      - description: Sort
        enum:
        - asc
        - desc
        in: query
        name: sort
        type: string
      - description: Page number pagination
        in: query
        name: page
        type: integer
      - description: Limit pagination
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my projects
      tags:
      - Me Project
  /api/me/schedules:
    get:
      consumes:
      - application/json
      operationId: my-schedule
      parameters:
      - description: Search by name
        in: query
        name: q
        type: string
      - description: Search by type
        enum:
        - all
        - meeting
        - discussion
        - review
        - presentation
        - etc
        in: query
        name: type
        type: string
      - description: Start date
        in: query
        name: start
        type: string
      - description: End date
        in: query
        name: end
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my schedule
      tags:
      - Me
  /api/me/task/count:
    get:
      consumes:
      - application/json
      operationId: my-task-count
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my task count
      tags:
      - Me Task
  /api/me/task/overview:
    get:
      consumes:
      - application/json
      operationId: my-task-overview
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my task overview
      tags:
      - Me Task
  /api/me/task/status:
    get:
      consumes:
      - application/json
      operationId: my-task-list-status
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my task list by status
      tags:
      - Me Task
  /api/me/tasks:
    get:
      consumes:
      - application/json
      operationId: my-tasks
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get my tasks
      tags:
      - Me Task
...
  /api/task/count:
    get:
      consumes:
      - application/json
      operationId: task-count
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task count
      tags:
      - Task
  /api/task/overview:
    get:
      consumes:
      - application/json
      operationId: task-overview
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task overview
      tags:
      - Task
  /api/task/status:
    get:
      consumes:
      - application/json
      operationId: task-list-status
      produces:
      - application/json
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: Get task list by status
      tags:
      - Task
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: Authorization
    type: apiKey
  BasicAuth:
    type: basic
swagger: "2.0"